Gunzip a file

gunzip a file

Usage

gunzip(
  filename,
  destname = gsub("[.]gz$", "", filename),
  overwrite = FALSE,
  remove = TRUE,
  BFR.SIZE = 1e+07
)

Arguments

filename

The filename to be unzipped

destname

The destination file

overwrite

Boolean indicating whether or not to overwrite a destfile of the same name

remove

Boolean indicating whether or not to remove the original file after completion

BFR.SIZE

The size of the read buffer....

Value

Invisibly, the number of bytes read.

Details

This function was stripped out of R.utils due to breaking some stuff on the bioconductor build machine.

See also

Author

Original author: Henrik Bengtsson

Examples

# gunzip('file.gz',remove=FALSE)