In some cases, instead of individual sample records (GSM) containing information regarding sample phenotypes, the GEO Series contains that information in an attached data table. And example is given by GSE3494 where there are two data tables with important information contained within them. Using getGEO with the standard parameters downloads the GSEMatrix file which, unfortunately, does not contain the information in the data tables. This function simply downloads the ``header'' information from the GSE record and parses out the data tables into R data.frames.

getGSEDataTables(GSE)

Arguments

GSE

The GSE identifier, such as ``GSE3494''.

Value

A list of data.frames.

References

http://www.ncbi.nlm.nih.gov/geo

See also

getGEO

Examples

dfl = getGSEDataTables("GSE3494")

No encoding supplied: defaulting to UTF-8.

lapply(dfl,head)

#> [[1]]
#>   INDEX (ID) p53 seq mut status (p53+=mutant; p53-=wt)
#> 1    X101B88                                      p53+
#> 2    X102B06                                      p53+
#> 3    X104B91                                      p53+
#> 4    X110B34                                      p53+
#> 5    X111B51                                      p53+
#> 6    X127B00                                      p53+
#>   p53 DLDA classifier result (0=wt-like, 1=mt-like) DLDA error (1=yes, 0=no)
#> 1                                                 1                        0
#> 2                                                 1                        0
#> 3                                                 0                        1
#> 4                                                 1                        0
#> 5                                                 1                        0
#> 6                                                 1                        0
#>   Elston histologic grade ER status PgR status age at diagnosis tumor size (mm)
#> 1                      G3       ER-       PgR-               40              12
#> 2                      G3       ER+       PgR+               51              26
#> 3                      G3       ER+       PgR+               80              24
#> 4                      G2       ER+       PgR+               74              20
#> 5                      G3       ER+       PgR+               41              33
#> 6                      G3       ER+       PgR+               57              22
#>   Lymph node status DSS TIME (Disease-Specific Survival Time in years)
#> 1               LN-                                             11.833
#> 2               LN-                                             11.833
#> 3               LN?                                              3.583
#> 4               LN-                                             11.667
#> 5               LN-                                              7.167
#> 6               LN-                                              4.667
#>   DSS EVENT (Disease-Specific Survival EVENT; 1=death from breast cancer, 0=alive or censored )
#> 1                                                                                             0
#> 2                                                                                             0
#> 3                                                                                             0
#> 4                                                                                             0
#> 5                                                                                             1
#> 6                                                                                             1
#> 
#> [[2]]
#>   GEO Sample Accession # Patient ID Affy platform
#> 1               GSM79114    X100B08      HG-U133A
#> 2               GSM79115    X101B88      HG-U133A
#> 3               GSM79116    X102B06      HG-U133A
#> 4               GSM79117    X103B41      HG-U133A
#> 5               GSM79118    X104B91      HG-U133A
#> 6               GSM79119    X105B13      HG-U133A
#> 
#> 

Author
Sean Davis <sdavis2@mail.nih.gov>